Replace the last (double) cast with (float)
The final energy demand fields (1.3.1) added a (double) cast after #179
had replaced the other eight, so 1.4.0 still raises E_DEPRECATED on php
8.5 for every realty that carries an energy pass with an eebwert.

(float) is the canonical spelling of the same cast, so behaviour is
unchanged.

+++ src/Justimmo/Model/Wrapper/V1/RealtyWrapper.php
@@ -444,7 +444,7 @@ class RealtyWrapper extends AbstractWrapper
                     ->setValidUntil($this->cast($energiepassXml->gueltig_bis, 'datetime'));
 
                 if (isset($energiepassXml->eebwert)) {
-                    $energiepass->setFinalEnergyDemandValue((double) $energiepassXml->eebwert);
+                    $energiepass->setFinalEnergyDemandValue((float) $energiepassXml->eebwert);
                 }
 
                 if (isset($energiepassXml->eebklasse)) {
